Fix QLTransactionTest.RemoteBootstrap test
Summary:
The following issues addressed:
1) When backoff waiter is configured with small max_wait attempts could grow very fast and cause overflow.
To fix this issue we will use max int32_t when attempt is big enough.

2) -k option was broken in repeat_unit_test.sh.

3) Infinite recursion happens when retryable task is trying to schedule during shutdown.
Added handling of this scenario.

4) Enabled coredump in ASAN builds.

Test Plan: ybd asan --cxx-test client_ql-transaction-test --gtest_filter QLTransactionTest.RemoteBootstrap -n 10

Reviewers: mikhail, robert, hector, timur

Reviewed By: timur

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D5165

+++ src/yb/util/backoff_waiter.h
@@ -29,7 +29,7 @@ class BackoffWaiter {
 
   bool Wait();
 
-  int attempt() const {
+  size_t attempt() const {
     return attempt_;
   }
 
@@ -39,7 +39,7 @@ class BackoffWaiter {
   }
 
  private:
-  int attempt_ = 0;
+  size_t attempt_ = 0;
   std::chrono::steady_clock::time_point deadline_;
   std::chrono::steady_clock::duration max_wait_;
 };
